Fix popup fade-in on Safari
Previously the opacity transition on Safari would correctly fade the
background however the text would just _appear_ at the end of the
transition. This appears to be some quirk of setting visibility where
the text doesn't adjust the visibility until the end of its transition.

To fix this it doesn't ever set itself invisible and just only
transitions into `hover` rather than transitioning both ways. However
this causes an issue where they are always selectable (hanging below
their name) and when you hover below the name the orcid pops up. But
there is no reason to be able to select the orcid-tooltip (as you select
the icon to open it) so it just disables `pointer-events`.