Adjust positioning of tooltip arrow
The positioning of the arrow was adjusted in 21bae97 to make it
cross-browser consistent however this shifted it a long way away from
the item being pointed to. This re-adjusts it so that the arrow points
directly at the 'iD' icon and appears to work identically across Chrome,
Firefox, Safari and Edge.