Prevent the tooltip text from being copied
As a quirk of removing `visibility: hidden` the tooltip text would now
be copied if you just try to select all of the names associated with an
EPrint. To solve this we have to set `user-select` (and its tagged
versions).

In a similar vein you now also can't select the 'iD' icon as this would
previously add '[ORCID logo]' to the copy-pasted text.